Refactor ChallengeMiniCard to use ViewChallengeButton component
