+++ src/components/badges/ChallengeCompletedBadge.tsx
@@ -1,0 +1,11 @@
+import { Badge } from "@mantine/core";
+
+export default function ChallengeCompletedBadge() {
+    return (
+        <div>
+            <Badge color="green" radius="sm">
+                Completed
+            </Badge>
+        </div>
+    );
+}
